Pinvoke x11 xrandr for linux resolutions (#107)
* Pinvoke x11 xrandr for linux resolutions
* Replace events with reactive
* Replace autofac with microsoft DI
* Enable warnings as errors

+++ Sims1WidescreenPatcher.Core/Models/CompositeImageJob.cs
@@ -19,7 +19,7 @@ public class CompositeImageJob : BaseImageProcessingJob
         }
         
         using var image = new MagickImage(ImageBytes!);
-        using var background = new MagickImage(new MagickColor(Color), Width, Height);
+        using var background = new MagickImage(new MagickColor(Color), (uint)Width, (uint)Height);
         background.Composite(image, Gravity.Center);
         SetCommonBmpSettings(background);
         background.Write(Output!);
